+++ packages/core/src/memory/index.ts
@@ -1,4 +1,12 @@
-import { AssistantContent, ToolContent, ToolResultPart, UserContent, Message as AiMessage } from 'ai';
+import {
+  AssistantContent,
+  ToolContent,
+  ToolResultPart,
+  UserContent,
+  Message as AiMessage,
+  CoreToolMessage,
+  ToolInvocation,
+} from 'ai';
 
 // Types for the memory system
 export type MessageType = {
@@ -27,6 +35,107 @@ export type ThreadType = {
  * conversation threads and messages.
  */
 export abstract class MastraMemory {
+  MAX_CONTEXT_TOKENS?: number;
+
+  estimateTokens(text: string): number {
+    return Math.ceil(text.split(' ').length * 1.3);
+  }
+
+  parseMessages(messages: MessageType[]): MessageType[] {
+    return messages.map(mssg => ({
+      ...mssg,
+      content: typeof mssg.content === 'string' ? JSON.parse((mssg as MessageType).content as string) : mssg.content,
+    }));
+  }
+
+  convertToUIMessages(messages: MessageType[]): AiMessage[] {
+    function addToolMessageToChat({
+      toolMessage,
+      messages,
+      toolResultContents,
+    }: {
+      toolMessage: CoreToolMessage;
+      messages: Array<AiMessage>;
+      toolResultContents: Array<ToolResultPart>;
+    }): { chatMessages: Array<AiMessage>; toolResultContents: Array<ToolResultPart> } {
+      const chatMessages = messages.map(message => {
+        if (message.toolInvocations) {
+          return {
+            ...message,
+            toolInvocations: message.toolInvocations.map(toolInvocation => {
+              const toolResult = toolMessage.content.find(tool => tool.toolCallId === toolInvocation.toolCallId);
+
+              if (toolResult) {
+                return {
+                  ...toolInvocation,
+                  state: 'result',
+                  result: toolResult.result,
+                };
+              }
+
+              return toolInvocation;
+            }),
+          };
+        }
+
+        return message;
+      }) as Array<AiMessage>;
+
+      const resultContents = [...toolResultContents, ...toolMessage.content];
+
+      return { chatMessages, toolResultContents: resultContents };
+    }
+
+    const { chatMessages } = messages.reduce(
+      (obj: { chatMessages: Array<AiMessage>; toolResultContents: Array<ToolResultPart> }, message) => {
+        if (message.role === 'tool') {
+          return addToolMessageToChat({
+            toolMessage: message as CoreToolMessage,
+            messages: obj.chatMessages,
+            toolResultContents: obj.toolResultContents,
+          });
+        }
+
+        let textContent = '';
+        let toolInvocations: Array<ToolInvocation> = [];
+
+        if (typeof message.content === 'string') {
+          textContent = message.content;
+        } else if (Array.isArray(message.content)) {
+          for (const content of message.content) {
+            if (content.type === 'text') {
+              textContent += content.text;
+            } else if (content.type === 'tool-call') {
+              const toolResult = obj.toolResultContents.find(tool => tool.toolCallId === content.toolCallId);
+              toolInvocations.push({
+                state: toolResult ? 'result' : 'call',
+                toolCallId: content.toolCallId,
+                toolName: content.toolName,
+                args: content.args,
+                result: toolResult?.result,
+              });
+            }
+          }
+        }
+
+        obj.chatMessages.push({
+          id: message.id,
+          role: message.role as AiMessage['role'],
+          content: textContent,
+          toolInvocations,
+        });
+
+        return obj;
+      },
+      { chatMessages: [], toolResultContents: [] } as {
+        chatMessages: Array<AiMessage>;
+        toolResultContents: Array<ToolResultPart>;
+      },
+    );
+
+    return chatMessages;
+  }
+
   /**
    * Retrieves a specific thread by its ID
    * @param threadId - The unique identifier of the thread
@@ -84,7 +193,7 @@ export abstract class MastraMemory {
    * @param toolName - The name of the tool that was called
    * @returns Promise resolving to the cached tool result or null if not found
    */
-  abstract getCachedToolResult({
+  abstract getToolResult({
     threadId,
     toolArgs,
     toolName,
@@ -96,10 +205,10 @@ export abstract class MastraMemory {
 
   /**
    * Checks if an un-expired tool call arg exists in a thread
-   * @param hashedToolCallArgs - The hashed tool call information (args, threadId, toolName) to check for
+   * @param hashedArgs - The hashed tool call information (args, threadId, toolName) to check for
    * @returns Promise resolving to true if the un-expired tool call arg exists, false otherwise
    */
-  abstract checkIfValidArgExists({ hashedToolCallArgs }: { hashedToolCallArgs: string }): Promise<boolean>;
+  abstract validateToolCallArgs({ hashedArgs }: { hashedArgs: string }): Promise<boolean>;
 
   /**
    * Helper method to create a new thread
